requirements.txt
python-dotenv==1.0.0
openai==1.12.0
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
jinja2==3.1.3
flask==3.0.2
flask-cors==4.0.0
asgiref==3.7.2
werkzeug==3.0.1
pathlib==1.0.1
requests==2.31.0
python-slugify==8.0.4
aiohttp==3.9.3
werkzeug==3.0.1
pathlib==1.0.1
requests==2.31.0
python-slugify==8.0.4  
aiohttp==3.9.3  